backend/requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==40.0.2
Django==4.1.9
sqlparse==0.4.4
django-cors-headers==3.14.0
django-unixtimestampfield==0.5.0
djangorestframework==3.14.0
pycparser==2.21
python-dotenv==1.0.0
PyGithub==1.58.1
pytz==2023.3
six==1.16.0
psycopg2-binary==2.9.6
django-extensions==3.2.3
